ec: Resolve build warnings from using gcc 14.2
GCC 14.2 appears to be more strict than previous versions when it comes
to type conversions and static casts.  Address build warnings.

BUG=b:380129441
TEST=zmake build brox-ish brox-ish-ec orisa orisa-ish

+++ zephyr/shim/src/watchdog.c
@@ -161,7 +161,7 @@ __maybe_unused static void wdt_warning_handler(const struct device *wdt_dev,
 #ifdef CONFIG_RISCV
 	exception_address = csr_read(mepc);
 	printk("WDT pre-warning MEPC:%p THREAD_NAME:%s\n",
-	       (void *)exception_address, thread_name);
+	       (void *)(long)exception_address, thread_name);
 #else
 	/* TODO(b/176523207): watchdog warning message */
 	printk("Watchdog deadline is close! THREAD_NAME:%s\n", thread_name);
